services/ingest/requirements.txt
requests
click
psycopg2-binary
tqdm